correções

+++ src/main/java/br/com/bluesoft/desafio/dtos/PrecoDto.java
@@ -1,16 +1,18 @@
 package br.com.bluesoft.desafio.dtos;
 
+import java.math.BigDecimal;
+
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 
 @JsonIgnoreProperties(ignoreUnknown = true)
 public class PrecoDto {
-	private float preco;
+	private BigDecimal preco;
 	private int quantidade_minima;
-	
-	public float getPreco() {
+
+	public BigDecimal getPreco() {
 		return preco;
 	}
-	public void setPreco(float preco) {
+	public void setPreco(BigDecimal preco) {
 		this.preco = preco;
 	}
 	public int getQuantidade_minima() {
@@ -22,5 +24,5 @@ public class PrecoDto {
 	@Override
 	public String toString() {
 		return "{preco=" + preco + ", quantidade_minima=" + quantidade_minima + "}";
-	}	
+	}
 }
